Add cardUtils
New file containing useful functions to handle cards
