+++ ai.c
@@ -1,6 +1,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include "core.h"
+#include "cardUtils.h"
 
 Card getAICardFirstAvailable(Card *cardArray, int nbOfCards) { //TEMPORARY FOR TEST PUROPOSES, WILL NEED AN UPDATE LATER
     Card chosenCard;
